Revert "머지 후 이벤트게시판 충돌 관련 이슈 수정"
