+++ src/club/club.repository.ts
@@ -22,11 +22,10 @@ export class ClubRepository {
 
 
   async getClubs() {
-    const data = await this.clubRepository.createQueryBuilder("page")
-    .leftJoinAndSelect("page.user", "user")
-    .orderBy("page.id", "DESC") //최신순(내림차순)
-    .getMany();
-
+    const data = await this.clubRepository.find({
+      where: { deletedAt: null },
+      select: ["id", "title", "maxMembers", "createdAt", "userId", "category"],
+    });
 
     return data;
   }
@@ -131,6 +130,34 @@ export class ClubRepository {
     await this.clubRepository.softDelete(clubId);
   }
 
+  //페이지네이션
+  async paginatedResults(page, term?: string) {
+    const take = 5;
+    const selectedData = await this.clubRepository
+      .createQueryBuilder("Clubs")
+      .leftJoinAndSelect("Clubs.user", "user")
+      .orderBy("Clubs.id", "DESC") //최신순(내림차순)
+      .getMany();
+
+    console.log(selectedData);
+
+    const totalDataCount = selectedData.length; //불러온 데이터 목록 수
+    const startIndex = (page - 1) * take;
+    const endIndex = page * take;
+
+    const slicedData = selectedData.slice(startIndex, endIndex); // 페이지당 조회할 데이터 묶음
+    const lastPage = Math.ceil(totalDataCount / take); //생성될 페이지 수
+
+    const unitSize = 3; // 페이지 묶음 단위 : 3개씩 < 1 2 3>  <4 5 6>
+    const numOfUnits = Math.floor((page - 1) / unitSize); //<1 2 3> 페이지는 0 번째 index
+    const unitStart = numOfUnits * unitSize + 1; //0번째 묶음의 시작은 1페이지, 1번째 묶음 시작은 4페이지...
+    const unitEnd = unitStart + (unitSize - 1); //0번째 묶음의 끝은 3페이지, 1번째 묶음 끝은 6페이지
+    const paginatedDemand = { page, slicedData, lastPage, unitStart, unitEnd };
+
+    return {
+      ...paginatedDemand,
+    };
+  }
   async createAbusing(clubId: number, userId: number) {
     const data = await this.abusingClubRepository.insert({
       clubId,
@@ -139,17 +166,4 @@ export class ClubRepository {
 
     return data;
   }
-
-
-  // // 여기서 확정된 참여인원 확인 - 클럽 멤버들 확인
-  // async viewClubMembers(userId: Number, clubId: Number) {
-  //   const currentClubMember = await this.clubMembersRepository.find({
-  //     where: {
-  //       clubId,
-  //       isAccepted: true,
-  //     },
-  //   });
-  //   return { currentClub, currentClubMember };
-  // }
-
 }
